* Add SLEPC build scripts
* Build/Process the packages directory before attempting to build a dependency list and
  have the dependency generator use that instead of the template directory.

This is due to having to process dependencies that are template based. Example:

  PACKAGE=<SLEPC_ALT>-mpich-clang
  DEPENDENCIES=(<PETSC_ALT> mpich clang)

(specific version of SLEPc depends on a specific version of PETSc)

Where as the actual <PETSC_ALT> module (aka: filename. see TODO) had not been processed by
the template engine and therefor the actual dependency is still unknown and according to
the dependency generator, a literal: '<PETSC_ALT>' being the dependency. Which of course is
never found. And therefor has no dependency, so it runs in an incorrect order and fails.

TODO:
Now that the above is in place, stop using the filename as the dependency descriptor. That was
wrong, a bad idea, and surprisingly worked for this long.