CCACHE when building the Package
Optimize the build system when building the package through the use of CCACHE.
Fix the discrepancy with package.tar and package.tar.gz in RPM SPEC file.
Pango requires harfbuzz to be built with glib.
No longer run the download_prerequisites script for GCC. It was prone to failure, and was really only intended to be run on Linux machines.
For real this time, remove the lapack/blas libs from trilinos pointing to petsc.
Give the clean up routine some much needed love.
Perform an MD5 checksum on downloaded files to better know when the system should download a new version.