Remove OpenMPI from package to save space
Remove OpenMPI from the package to prevent RPM package limitation
bug (https://bugs.launchpad.net/rpm/+bug/651439).

Allow miniconda to install boost.

Alter SPEC file for RPM packages to be a bit more robust.

Address a rather nasty bug in the functions bash script which
had the potential of deleting the entire contents of
$PACKAGE_DIR, if $PACKAGE and $BASE_NAME are not set (brought on
by the alteration of the boost module).

This is a merge of several PRs. The reason for this, is that we had
an unfortunate timing issue of PRs hit the build machines. The
easiest (and least time consuming) fix, is to combined all these
PRs into a single PR.

Closes #126 #123 #121