Update LLVM to 5.0.1 add LLDB
Update to LLVM 5.0.1

Build a minimal clang target so we can use that to
build the entire LLVM project (OpenMP, compier-rt,
lldb, etc). LLDB requires itself to be built using the
same compiler we are building. So it made sense to
adopt a bootstrapping method.

Add swig, as that is a new requirement when building
LLDB.

Closes #132