The bootstrapper needs to be built in shared mode
The bootstrapper was creating a set of libs/bins at the 18GB
level. No good. Build in shared mode.

I found doing so, I had to then alter all the relative path
links to be absolute ones once the build was done.

Not sure why I can not do the same for the final target. Need
to look into this. I'd like to test a shared library version
of LLVM/Clang.