Use a tempfile instead of subprocess buffer
If the output of subprocess if large enough, it will hang and wait for someone to communicate with
it (to empty the buffer). This is was what was happening on our Fedora 28 VM, causing it to hang.
(https://docs.python.org/2/library/subprocess.html#popen-objects)

Syntax bugs was creating an empty plist file in our icecream client script.