Remove SEACAS tools from moose-environment package
Remove seacas from moose-environment. This project no longer builds correctly for the platform and compilers we
are providing. The issue, is that seacas does not build correctly with recent GCC versions and modern versions
of GCC is what we need, when building modern versions of LLVM.

We will have to maintain a separate package for SEACAS Tools, which coincidentally, will need to contain its own
GCC stack (5.2.0 as I recall).

Machines where python3 is the default python interpreter, we need to execute python2 when invoking the configure
scripts for PETSc.