Allow a separate intel stack
Allow the package builder to build a library stack based on an Intel compiler.

  --with-intel64 (defaults to /opt/intel)
or
  --with-intel64 /root/path/to/intel

The naming convention will append '_intel' to what ever the prefix is at the time:

  --prefix /opt/moose --with-intel64

...will produce the normal GNU library stack in /opt/moose, in addition to an Intel
one located at: /opt/moose_intel. This is so we can separate the GNU stack from an
Intel one, when we go to create the moose-environment redistributable.