Remove --enable-fast for mpich-clang
I am seeing errors during runtime with parallel tests. The backtraces point to
libpmpi.12.dylib. Much like this post:

   https://lists.mpich.org/pipermail/discuss/2016-July/004824.html

This PR is just a test to see if the newly added option (because I was doing it
wrong for the past 3 years), is the cause.

Added a message to be displayed about the users environment being changed (Macs).