Fix lcov, clang's ld, and urls.txt
Latest Lcov includes a makefile now. Use that.

Append path to llvm's bin directory instead of pre-prending when loading the the GCC module. This is necessary due in part to the addition of clang's linker we now build.

Fix that pesky issue when urls.txt is missing, conda create environment fails, unless you own the conda installation.