Fix more Python2-3 issues
Turns out mumps is failing with missing libraries (mpiseq). Not sure why this
is just now cropping up. The only difference between the last successful build
is Python3 (and I suppose GCC8 to 9. But the failures are with the LLVM compiled
libraries, and that has not changed).

Removing the call to python2. Use environment defaults instead.