I can not codesign lldb-8.0.1. Update to 9.
There is something wrong with trying to codesign LLDB on 8.0.x.
Lets upgrade to 9, as that appears to work on my workstation.

LLVM _still_ links to rapth Python, which in the end, is not the
python library I am asking LLDB to link to. It is still necessary
to use install_name_tool to fix these issues.

Use the correct codesign certificate path in the Keychain.